requirements.txt
gTTS==2.3.2
RPi.GPIO==0.7.0
numpy==1.21.6
pydub==0.25.1
PyAudio==0.2.11
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-speech==2.19.0
googleapis-common-protos==1.59.0
SpeechRecognition==3.9.0
